requirements.txt
annotated-types==0.6.0
blis==0.7.11
brotlipy==0.7.0
catalogue==2.0.10
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.7
clip==1.0
cloudpathlib==0.16.0
confection==0.1.4
cryptography==41.0.3
cymem==2.0.8
deepspeed==0.4.0
filelock==3.13.1
fsspec==2023.10.0
ftfy==6.1.1
huggingface-hub==0.18.0
idna==3.4
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.4
mkl-fft==1.3.0
mkl-random==1.0.2
mkl-service==2.3.0
murmurhash==1.0.10
ninja==1.11.1.1
numpy==1.26.1
olefile==0.46
packaging==23.2
pandas==2.1.4
Pillow==8.3.1
pip==23.3
preshed==3.0.9
protobuf==4.25.0
psutil==5.9.6
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pyOpenSSL==23.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.2
scipy==1.11.3
setuptools==68.0.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tensorboardX==1.8
thinc==8.2.2
threadpoolctl==3.2.0
timm==0.9.8
torch==1.11.0+cu115
torchaudio==0.9.0a0+a85b239
torchsummary==1.5.1
torchvision==0.10.1
tqdm==4.66.1
triton==2.1.0
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
wcwidth==0.2.8
weasel==0.3.4
wheel==0.41.2
